feat: create abstraction for boolean and number fields
