--- packages/twenty-front/src/modules/workflow/components/WorkflowEditActionFormField.tsx
@@ -1,19 +1,0 @@
-import { FormNumberFieldInput } from '@/object-record/record-field/form-types/components/FormNumberFieldInput';
-
-type WorkflowEditActionFormFieldProps = {
-  defaultValue: string;
-};
-
-export const WorkflowEditActionFormField = ({
-  defaultValue,
-}: WorkflowEditActionFormFieldProps) => {
-  return (
-    <FormNumberFieldInput
-      defaultValue={defaultValue}
-      placeholder="Placeholder"
-      onPersist={(value) => {
-        console.log('save value to database', value);
-      }}
-    />
-  );
-};
